log with toString()
